# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dragon Ball Rest Api",
        "description": "Restful api to provide information about the japanese anime",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Kaiosama",
            "url": "http://www.noonecares.com/support",
            "email": "support@noonecares.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "https://localhost:{port}/{basePath}",
            "description": "development",
            "variables": {
                "port": {
                    "enum": [ "3000" ],
                    "default": "3000"
                },
                "basePath": {
                    "default": "api"
                }
            }
        }
    ],
    "components": {
        "securitySchemes": {
          "ApiKeyAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "X-ACCESS-TOKEN"
          }
        },
        "schemas": {
            "User": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "profile": {
                  "type": "integer"
                }
              }
            },
            "Authenticated": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "Character": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "avatar": {
                  "type": "string"
                }
              }
            },
            "Saga": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                }
            },
            "Film": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "movies": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "cover": {
                  "type": "string"
                }
              }
            },
            "Created": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                } 
              }
            },
            "Updated": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                } 
              }
            },
            "Deleted": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                } 
              }
            },
            "Error": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "message": {
                  "type": "string"
                }
              }
            }
        }
    },
    "paths": {
        "/auth": {
            "post": {
              "tags": ["Authentication"],
              "description": "This endpoint allows the user to authenticate and obtain interaction token",
              "operationId": "authenticate",
              "consumes": [
                "application/x-www-form-urlencoded"
              ],
              "produces": [
                "application/xml",
                "application/json"
              ],
              "requestBody": {
                  "description": "Username and password to get authenticated",
                  "required": true,
                  "content": {
                    "application/x-www-form-urlencoded": {
                      "schema": {
                        "$ref": "#/definitions/User"
                      }
                    }
                  }
              },
              "responses": {
                "200": {
                  "description": "Ok",
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/Authenticated"
                      }
                    }
                  }
                }
              }
            }
        },
        "/characters": {
          "get": {
            "tags": ["Characters"],
            "description": "Get characters info",
            "operationId": "get_characters",
            "parameters": [],
            "responses": {
              "200": {
                "description": "Ok",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Character"
                    }
                  }
                }
              },
              "404": {
                  "description": "The specified resource was not found",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#components/schemas/Error"
                          }
                      }
                  }
              },
              "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#components/schemas/Error"
                        }
                    }
                }
              }
            }
          },
          "post": {
            "tags": ["Characters"],
            "description": "Create a character record",
            "operationId": "create_character",
            "consumes": [
              "application/x-www-form-urlencoded"
            ],
            "produces": [
              "application/xml",
              "application/json"
            ],
            "requestBody": {
                "description": "Character info",
                "required": true,
                "content": {
                  "application/x-www-form-urlencoded": {
                    "schema": {
                      "$ref": "#components/schemas/Character"
                    }
                  }
                }
            },
            "responses": {
              "201": {
                "description": "Resource created",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Created"
                    }
                  }
                }
              }
            },
            "security": [
              {
                "ApiKeyAuth": []
              }
            ]
          }
        },
        "/characters/{id}": {
          "delete": {
            "tags": ["Characters"],
            "description": "Deleting a character",
            "operationId": "delete_character",
            "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the character",
                  "required": true,
                  "schema": {
                      "type": "string"
                  },
                  "style": "simple"
              }
            ],
            "consumes": [
              "application/x-www-form-urlencoded"
            ],
            "produces": [
              "application/xml",
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "Ok",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Deleted"
                    }
                  }
                }
              },
              "404": {
                  "description": "The specified resource was not found",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#components/schemas/Error"
                          }
                      }
                  }
              },
              "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#components/schemas/Error"
                        }
                    }
                }
              }
            },
            "security": [
              {
                "ApiKeyAuth": []
              }
            ]
          },
          "patch": {
            "tags": ["Characters"],
            "description": "Updates a character",
            "operationId": "update_character",
            "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the character",
                  "required": true,
                  "schema": {
                      "type": "string"
                  },
                  "style": "simple"
              }
            ],
            "consumes": [
              "application/x-www-form-urlencoded"
            ],
            "produces": [
              "application/xml",
              "application/json"
            ],
            "requestBody": {
              "description": "Character info",
              "required": true,
              "content": {
                "application/x-www-form-urlencoded": {
                  "schema": {
                    "$ref": "#components/schemas/Character"
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Resource updated",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Updated"
                    }
                  }
                }
              }
            },
            "security": [
              {
                "ApiKeyAuth": []
              }
            ]
          }
        },
        "/characters/{name}": {
            "get": {
                "tags": ["Characters"],
                "description": "Get a single character info",
                "operationId": "get_character",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the character being searched",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                          "application/json": {
                            "schema": {
                              "$ref": "#/components/schemas/Character"
                            }
                          }
                        }
                    }
                }
            }
        },
        "/sagas": {
            "get": {
                "tags": ["Sagas"],
                "description": "Get sagas info",
                "operationId": "get_sagas",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Saga"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
              "tags": ["Sagas"],
              "description": "Create a new saga",
              "operationId": "create_sagas",
              "parameters": [],
              "requestBody": {
                "description": "Saga info",
                "required": true,
                "content": {
                  "application/x-www-form-urlencoded": {
                    "schema": {
                      "$ref": "#components/schemas/Saga"
                    }
                  }
                }
              },
              "responses": {
                  "201": {
                    "description": "Resource created",
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/Created"
                        }
                      }
                    }
                  }
              },
              "security": [
                {
                  "ApiKeyAuth": []
                }
              ]
          }
        },
        "/sagas/{name}": {
            "get": {
                "tags": ["Sagas"],
                "description": "Get a single saga info",
                "operationId": "get_saga",
                "parameters": [
                  {
                    "name": "name",
                    "in": "path",
                    "description": "Name of the saga being searched",
                    "required": true,
                    "schema": {
                        "type": "string"
                    },
                    "style": "simple"
                  }
                ]
            }
        },
        "/sagas/{id}": {
          "delete": {
            "tags": ["Sagas"],
            "description": "Deleting a saga",
            "operationId": "delete_saga",
            "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the saga",
                  "required": true,
                  "schema": {
                      "type": "string"
                  },
                  "style": "simple"
              }
            ],
            "consumes": [
              "application/x-www-form-urlencoded"
            ],
            "produces": [
              "application/xml",
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "Ok",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Deleted"
                    }
                  }
                }
              },
              "404": {
                  "description": "The specified resource was not found",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#components/schemas/Error"
                          }
                      }
                  }
              },
              "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#components/schemas/Error"
                        }
                    }
                }
              }
            },
            "security": [
              {
                "ApiKeyAuth": []
              }
            ]
          }
        },
        "/films": {
          "get": {
            "tags": ["Films"],
            "description": "Get a list of films",
            "operationId": "get_films",
            "responses": {
              "200": {
                  "description": "Ok",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Film"
                          }
                      }
                  }
              }
            }
          },
          "post": {
            "tags": ["Films"],
              "description": "Create a new film",
              "operationId": "create_film",
              "parameters": [],
              "requestBody": {
                "description": "Film info",
                "required": true,
                "content": {
                  "application/x-www-form-urlencoded": {
                    "schema": {
                      "$ref": "#components/schemas/Film"
                    }
                  }
                }
              },
              "responses": {
                  "201": {
                    "description": "Resource created",
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/Created"
                        }
                      }
                    }
                  }
              },
              "security": [
                {
                  "ApiKeyAuth": []
                }
              ]
          }
        },
        "/films/{name}": {
          "get": {
              "tags": ["Films"],
              "description": "Get a single film info",
              "operationId": "get_film",
              "parameters": [
                  {
                      "name": "name",
                      "in": "path",
                      "description": "Name of the film being searched",
                      "required": true,
                      "schema": {
                          "type": "string"
                      },
                      "style": "simple"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Ok",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/Film"
                          }
                        }
                      }
                  }
              }
          }
        },
        "/films/{id}": {
          "delete": {
            "tags": ["Films"],
            "description": "Deleting a film",
            "operationId": "delete_film",
            "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the film",
                  "required": true,
                  "schema": {
                      "type": "string"
                  },
                  "style": "simple"
              }
            ],
            "consumes": [
              "application/x-www-form-urlencoded"
            ],
            "produces": [
              "application/xml",
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "Ok",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Deleted"
                    }
                  }
                }
              },
              "404": {
                  "description": "The specified resource was not found",
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#components/schemas/Error"
                          }
                      }
                  }
              },
              "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#components/schemas/Error"
                        }
                    }
                }
              }
            },
            "security": [
              {
                "ApiKeyAuth": []
              }
            ]
          }
        }
        
      },
      "securityDefinitions": {
        "api_key": {
          "type": "apiKey",
          "name": "api_key",
          "in": "header"
        }
      },
      "definitions": {
        "User": {
          "type": "object",
          "properties": {
            "user": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        }
      }
  }